INITIAL COMMIT:
This is the first section explaining how https or overall how secure connection is established.

run cmd "node SERVER" to start the server to listen on port 50001.
run cmd "node CLIENT" to initiate a seriese of http request.

follow the seriese of request of and responses.
this explains how a client secret is exchanged to the server,
with the help of server secret key.

This project also explains how a symmetric and asymmetric encryption works under the hood.
check enc.js file to see the maths behind generation of private and public key.

Disclaimer: This is just a sample project for education, in actuality things are way different.